pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel", "setuptools_scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bisontitan"
dynamic = ["version"]
description = "A robust, modular security suite for defensive security operations"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "BisonTitan Team"}
]
keywords = [
    "security", "malware", "scanner", "network", "forensics",
    "vulnerability", "pentest", "defensive", "windows", "siem"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Networking :: Monitoring",
]

# =============================================================================
# Core Dependencies (always installed)
# =============================================================================
dependencies = [
    # CLI & TUI
    "click>=8.1.0",
    "rich>=13.7.0",
    "textual>=0.47.0",

    # System utilities
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # Data processing
    "pandas>=2.1.0",
    "requests>=2.31.0",

    # Network analysis
    "scapy>=2.5.0",

    # Scheduling
    "schedule>=1.2.0",
]

# =============================================================================
# Optional Dependencies
# =============================================================================
[project.optional-dependencies]

# Malware scanning with YARA rules
yara = ["yara-python>=4.3.0"]

# Network scanning with nmap
nmap = ["python-nmap>=0.7.1"]

# SMB/Protocol analysis
protocols = ["impacket>=0.11.0"]

# Browser fingerprinting
browser = ["playwright>=1.40.0"]

# Windows-specific (event logs)
windows = ["pywin32>=306"]

# GeoIP lookups
geoip = [
    "geoip2>=4.7.0",
    "geocoder>=1.38.1",
]

# Web GUI dashboard
gui = [
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
]

# AI/LLM integration (optional)
ai = [
    "langchain-anthropic>=0.1.0",
    "langchain-core>=0.1.0",
]

# Development tools
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "pyinstaller>=6.0.0",
]

# Full installation (everything except AI)
full = [
    "bisontitan[yara,nmap,protocols,browser,windows,geoip,gui]",
]

# All features including dev
all = [
    "bisontitan[full,dev]",
]

# =============================================================================
# Entry Points
# =============================================================================
[project.scripts]
bisontitan = "bisontitan.cli:main"
bt = "bisontitan.cli:main"

[project.gui-scripts]
bisontitan-gui = "bisontitan.gui.app:main"

# =============================================================================
# Project URLs
# =============================================================================
[project.urls]
Homepage = "https://github.com/bisontitan/bisontitan"
Documentation = "https://github.com/bisontitan/bisontitan#readme"
Repository = "https://github.com/bisontitan/bisontitan"
Issues = "https://github.com/bisontitan/bisontitan/issues"
Changelog = "https://github.com/bisontitan/bisontitan/blob/main/CHANGELOG.md"

# =============================================================================
# Setuptools Configuration
# =============================================================================
# =============================================================================
# Setuptools SCM - Git-based versioning
# =============================================================================
[tool.setuptools_scm]
# Write version to file for runtime access
write_to = "src/bisontitan/_version.py"
write_to_template = '''"""Auto-generated version file. DO NOT EDIT."""
__version__ = "{version}"
__version_tuple__ = {version_tuple}
'''

# Version format
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"

# Fallback version if not in git repo
fallback_version = "1.0.0"

[tool.setuptools.packages.find]
where = ["src"]
include = ["bisontitan*"]

[tool.setuptools.package-data]
bisontitan = [
    "py.typed",
    "gui/*.py",
    "rules/*.yar",
    "rules/*.yara",
    "data/*.json",
    "data/*.yaml",
]

[tool.setuptools.exclude-package-data]
"*" = ["*.pyc", "__pycache__", "*.pyo"]

# =============================================================================
# Testing Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src/bisontitan"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# =============================================================================
# Code Quality Tools
# =============================================================================
[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "S",   # flake8-bandit (security)
]
ignore = [
    "S101",  # assert allowed in tests
    "S603",  # subprocess allowed
    "S607",  # partial path allowed
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "S106"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
disallow_untyped_defs = false
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "scapy.*",
    "yara.*",
    "nmap.*",
    "impacket.*",
    "win32.*",
    "pywintypes.*",
]
ignore_missing_imports = true
